requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.4.0
inflection==0.5.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.129
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.3
numpy==1.26.4
openai==1.50.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-alephalpha==0.31.1
opentelemetry-instrumentation-anthropic==0.31.1
opentelemetry-instrumentation-bedrock==0.31.1
opentelemetry-instrumentation-chromadb==0.31.1
opentelemetry-instrumentation-cohere==0.31.1
opentelemetry-instrumentation-google-generativeai==0.31.1
opentelemetry-instrumentation-groq==0.31.1
opentelemetry-instrumentation-haystack==0.31.1
opentelemetry-instrumentation-lancedb==0.31.1
opentelemetry-instrumentation-langchain==0.31.1
opentelemetry-instrumentation-llamaindex==0.31.1
opentelemetry-instrumentation-marqo==0.31.1
opentelemetry-instrumentation-milvus==0.31.1
opentelemetry-instrumentation-mistralai==0.31.1
opentelemetry-instrumentation-ollama==0.31.1
opentelemetry-instrumentation-openai==0.31.1
opentelemetry-instrumentation-pinecone==0.31.1
opentelemetry-instrumentation-qdrant==0.31.1
opentelemetry-instrumentation-replicate==0.31.1
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-threading==0.48b0
opentelemetry-instrumentation-together==0.31.1
opentelemetry-instrumentation-transformers==0.31.1
opentelemetry-instrumentation-urllib3==0.48b0
opentelemetry-instrumentation-vertexai==0.31.1
opentelemetry-instrumentation-watsonx==0.31.1
opentelemetry-instrumentation-weaviate==0.31.1
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-semantic-conventions-ai==0.4.1
opentelemetry-util-http==0.48b0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
posthog==3.6.6
protobuf==4.25.5
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
streamlit==1.38.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traceloop-sdk==0.31.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.13.0
zipp==3.20.2
